social_app/requirements.txt
Django==5.1
djangorestframework==3.15.2
psycopg2==2.9.9